Changed argument check to IP-focus
The previous check of the argument passed to the script focused on the case where a user doesn't provide input. There are lots of possible entry options that wouldn't be valid, so the logic has been reversed; it now checks if the entry is 3 numeric values separated by dots, and rejects anything else.

The regex is a relatively simple one, so it doesn't check for values outside the 0-255 range (or 1-254 range for the first octet).